skills/godogen/tools/requirements.txt
google-genai
requests
numpy
pillow
rembg
onnxruntime-gpu
nvidia-cudnn-cu12==9.*
